Fixed operator<<

+++ main.cpp
@@ -4,15 +4,15 @@
 using namespace std;
 
 int main(int argc, char *argv[]) {
-    vector<vector<double>> vec = {{5, -2, 2, 7},
+    vector<vector<float>> vec = {{5, -2, 2, 7},
                     {1, 0, 0, 3},
                     {-3, 1, 5, 0},
                     {3, -1, -9, 4}} ;
     
-    Matrix<double> matrixA(vec);
+    Matrix<float> matrixA(vec);
     cout << "matrixA: \n" << matrixA << endl ;
 
-    Matrix<double> matrixB(4,4);
+    Matrix<float> matrixB(4,4);
 
     
     for(unsigned int i= 0 ; i < matrixB.getRowCount(); i++)
@@ -26,7 +26,7 @@ int main(int argc, char *argv[]) {
     cout << matrixB << endl ;
 
     try {
-        Matrix<double> matrixC = matrixA + matrixB ;
+        Matrix<float> matrixC = matrixA + matrixB ;
         cout << "MATRIX : C = A + B" << endl ;
         cout << matrixC << endl ;
     }catch (const MatrixException &e){
@@ -34,7 +34,7 @@ int main(int argc, char *argv[]) {
     }
 
     try {
-        Matrix<double> matrixD = matrixA * matrixB ;
+        Matrix<float> matrixD = matrixA * matrixB ;
         cout << "MATRIX : D = A * B" << endl ;
         cout << matrixD << endl ;
     }catch (const MatrixException &e){
@@ -42,7 +42,7 @@ int main(int argc, char *argv[]) {
     }
 
     try {
-        Matrix<double> matrixAdjoint =  matrixA.adjoint();
+        Matrix<float> matrixAdjoint =  matrixA.adjoint();
         cout << "MATRIX : ADJOINT  A: " << endl ;
         cout << matrixAdjoint << endl ;
     }catch (const MatrixException &e){
@@ -52,11 +52,11 @@ int main(int argc, char *argv[]) {
     cout << "MATRIX : A DETERMINANT : " << matrixA.determinant() << endl ;
 
     try {
-        Matrix<double> matrixInv =  matrixA.inverse();
+        Matrix<float> matrixInv =  matrixA.inverse();
         cout << "MATRIX : A INVERSE :" << endl ;
         cout << matrixInv << endl ;
 
-        Matrix<double> matrixIdentity =  matrixA * matrixInv ;      
+        Matrix<float> matrixIdentity =  matrixA * matrixInv ;      
         cout << "MATRIX : A IDENTITY :" << endl ;
         cout << matrixIdentity << endl ;
     }catch (const MatrixException &e){
